Fix cryptohome for gcc 4.7 built.
Added explict converion in initialization list.

TEST=Manually built using gcc 4.7 for lumpy.
BUG=None

Reviewed-on: https://gerrit.chromium.org/gerrit/30771
Tested-by: Han Shen <shenhan@chromium.org>
Commit-Ready: Han Shen <shenhan@chromium.org>

+++ cryptohome/attestation.cc
@@ -459,11 +459,12 @@ bool Attestation::VerifyQuote(const SecureBlob& aik_public_key,
 
   // Check that the quoted value matches the given PCR value. We can verify this
   // by reconstructing the TPM_PCR_COMPOSITE structure the TPM would create.
-  const uint8_t header[] = {static_cast<uint8_t>(0), static_cast<uint8_t>(2),
-                            static_cast<uint8_t>(1), static_cast<uint8_t>(0),
-                            static_cast<uint8_t>(0), static_cast<uint8_t>(0),
-                            static_cast<uint8_t>(0),
-                            quote.quoted_pcr_value().size()};
+  const uint8_t header[] = {
+    static_cast<uint8_t>(0), static_cast<uint8_t>(2),
+    static_cast<uint8_t>(1), static_cast<uint8_t>(0),
+    static_cast<uint8_t>(0), static_cast<uint8_t>(0),
+    static_cast<uint8_t>(0),
+    static_cast<uint8_t>(quote.quoted_pcr_value().size())};
   SecureBlob pcr_composite = SecureCat(
       SecureBlob(header, arraysize(header)),
       ConvertStringToBlob(quote.quoted_pcr_value()));
